SUMMARY:  Fixing some compiler warnings on Kallikrates (Windows).
     BUGZILLA BUG #(s):

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  n/a

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  n/a

     3)  Did you modify or add any source or build system files?  no


     DESCRIPTION OF CHANGES:

     All mpi/serial tests pass on my workstation.

(checklist version 2009-07-14)

+++ packages/ml/src/Coarsen/ml_agg_ParMETIS.c
@@ -681,12 +681,12 @@ int ML_DecomposeGraph_with_ParMETIS( ML_Operator *Amatrix,
        each subdomain for each balance constraint */
       
     for( i=0 ; i<N_parts ; i++ )
-      tpwgts[i] = 1.0/N_parts;
+      tpwgts[i] = (float) (1.0/N_parts);
       
     /* imbalance tolerance for each vertex weight, 1 is the perfect
        balance and N_parts being perfect imbalance */
       
-    ubvec = 1.05;
+    ubvec = (float) 1.05;
       
   }
